+++ test/MiningOptimizationTraining/DataModel/ReturnArrivalData.h
@@ -31,7 +31,7 @@ namespace MiningOptimizationTraining
      */
     struct ReturnArrivalData
     {
-        uint16_t packed = UINT8_MAX;
+        uint16_t packed = UINT16_MAX;
         PositionAndVelocity nextPathStartPosition;
 
         // The number of frames to arrival at the target
